Extract repositories to insights module folder
