+++ packages/cli/src/modules/insights/repositories/insights-raw.repository.ts
@@ -1,7 +1,7 @@
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
 
-import { InsightsRaw } from '../../modules/insights/entities/insights-raw';
+import { InsightsRaw } from '../entities/insights-raw';
 
 @Service()
 export class InsightsRawRepository extends Repository<InsightsRaw> {
